refactor: move game states to hook useGameState
